feat: WorkbookCustomPage parameter 조건 추가 전체 필터 선택시 파라미터에 추가하지 않도록 조건추가

+++ src/main/java/com/example/back/question/domain/QuestionRepositoryCustom.java
@@ -7,6 +7,6 @@ import com.querydsl.core.types.OrderSpecifier;
 import java.util.List;
 
 public interface QuestionRepositoryCustom {
-    List<Question> searchQuestion(byte grade, byte month, int point, int page, int size, OrderSpecifier orderSpecifier);
+    List<Question> searchQuestion(int grade, int month, int point, int page, int size, OrderSpecifier orderSpecifier);
     List<Question> searchQuestionByRange(RangeRequest request);
 }
